refactor: Move restoreAccount logic into ui-keyring
