+++ packages/app-accounts/src/Restore.tsx
@@ -8,9 +8,8 @@ import { I18nProps } from '@polkadot/ui-app/types';
 import React from 'react';
 import { AddressSummary, Button, InputFile, Password } from '@polkadot/ui-app/index';
 import { InputAddress } from '@polkadot/ui-app/InputAddress';
-import createPair from '@polkadot/keyring/pair';
 import { decodeAddress } from '@polkadot/keyring';
-import { hexToU8a, isHex, isObject, u8aToString } from '@polkadot/util';
+import { isHex, isObject, u8aToString } from '@polkadot/util';
 import keyring from '@polkadot/ui-keyring/index';
 
 import translate from './translate';
@@ -121,6 +120,7 @@ class Restore extends React.PureComponent<Props, State> {
         isFileValid: false,
         json: null
       });
+      console.error(error);
     }
   }
 
@@ -140,19 +140,7 @@ class Restore extends React.PureComponent<Props, State> {
     }
 
     try {
-      const pair = createPair(
-        {
-          publicKey: decodeAddress(json.address),
-          secretKey: new Uint8Array()
-        },
-        json.meta,
-        hexToU8a(json.encoded)
-      );
-
-      // unlock, save account and then lock (locking cleans secretKey, so needs to be last)
-      pair.decodePkcs8(password);
-      keyring.addAccountPair(pair, password);
-      pair.lock();
+      const pair = keyring.restoreAccount(json, password);
 
       InputAddress.setLastValue('account', pair.address());
       onRestoreAccount();
